Introduce EnhancedList component
Refactor enhanceList function into EnhancedList component
