+++ console-frontend/src/app/resources/navigations/list.js
@@ -1,10 +1,8 @@
 import BlankStateTemplate from 'shared/blank-state'
-import enhanceList from 'ext/recompose/enhance-list'
 import React from 'react'
 import routes from 'app/routes'
-import { ActiveColumn, Avatar, columns, TableCell, Text } from 'shared/table-elements'
+import { ActiveColumn, Avatar, columns, EnhancedList, TableCell, Text } from 'shared/table-elements'
 import { apiNavigationDestroy, apiNavigationDuplicate, apiNavigationList } from 'utils'
-import { compose } from 'recompose'
 
 const BlankState = () => (
   <BlankStateTemplate
@@ -34,16 +32,19 @@ const NavigationsRow = ({ record, highlightInactive }) => (
   </React.Fragment>
 )
 
-export default compose(
-  enhanceList({
-    title: 'Navigations',
-    columns,
-    defaultSorting: { column: 'active', direction: 'asc' },
-    blankState: BlankState,
-    buttonText: 'Create new',
-    api: { fetch: apiNavigationList, destroy: apiNavigationDestroy, duplicate: apiNavigationDuplicate },
-    routes: { create: routes.navigationCreate, edit: routes.navigationEdit },
-    help: location => ({ single: true, stepName: 'navigations', stageName: 'initial', pathname: location.pathname }),
-    highlightInactive: ['triggerIds'],
-  })
-)(NavigationsRow)
+const NavigationsList = () => (
+  <EnhancedList
+    api={{ fetch: apiNavigationList, destroy: apiNavigationDestroy, duplicate: apiNavigationDuplicate }}
+    BlankState={BlankState}
+    buttonText="Create new"
+    columns={columns}
+    defaultSorting={{ column: 'active', direction: 'asc' }}
+    helpStep="navigations"
+    highlightInactive={['triggerIds']}
+    ResourceRow={NavigationsRow}
+    routes={{ create: routes.navigationCreate, edit: routes.navigationEdit }}
+    title="Navigations"
+  />
+)
+
+export default NavigationsList
