+++ console-frontend/src/shared/table-elements/index.js
@@ -1,6 +1,7 @@
 import ActiveColumn from './active-column'
 import Avatar from './avatar'
 import columns from './columns'
+import EnhancedList from './enhanced-list'
 import Picture from './picture'
 import TableCell from './table-cell'
 import TableHead from './table-head'
@@ -8,4 +9,4 @@ import TableRow from './table-row'
 import TableToolbar from './table-toolbar'
 import Text from './text'
 
-export { ActiveColumn, Avatar, columns, Picture, TableCell, TableHead, TableRow, TableToolbar, Text }
+export { ActiveColumn, Avatar, columns, EnhancedList, Picture, TableCell, TableHead, TableRow, TableToolbar, Text }
